Fix incorrect include paths in schedcp-cli
The create_and_run feature in the mcp directory
(schedcp-cli create-and-run /path/to/my_scheduler.bpf.c)
failed to compile due to incorrect include path configuration,
which caused clang to be unable to locate vmlinux.h.

This patch corrects the include paths, allowing vmlinux.h to be
found properly and restoring the expected functionality of the
create-and-run command.

Signed-off-by: YU-WEI,HSU <blackfire106.1231@gmail.com>